Swinging Added
Got the basic function for smooth swinging implemented. I need to do
some work on the movement to get them to work properly and make the
swing generalized for the all situations of holding and releasing. Also,
I had a bunch of issues running the movement stuff due to nonexistant
axis. Will look more into this tomorrow.